+++ Products/PluggableAuthService/interfaces/events.py
@@ -42,3 +42,8 @@ class IPropertiesUpdatedEvent(IPASEvent):
     properties = Attribute('List of modified property ids')
 
 
+class IGroupDeletedEvent(IPASEvent):
+    """A group has been removed.
+    """
+
+