+++ CHANGES.rst
@@ -4,8 +4,11 @@ Change Log
 1.10.1 (unreleased)
 -------------------
 
+- Add new event to be able to notify group deletion.
+
 - Fix usage of os.path.split(). Could result in Errors during import
   on Windows.
+
 
 1.10.0 (2013-02-19)
 -------------------